move pretty printer to its own module
